blah
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@323 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/unrrdu/unu.c
@@ -138,7 +138,7 @@ usage(char *me) {
   }
 
   sprintf(buff, "--- %s: command-line nrrd interface ---", me);
-  sprintf(fmt, "%%%ds\n", (80-strlen(buff))/2 + strlen(buff) - 1);
+  sprintf(fmt, "%%%ds\n", (int)((80-strlen(buff))/2 + strlen(buff) - 1));
   fprintf(stderr, fmt, buff);
   
   for (i=0; cmdList[i]; i++) {
